First Commit to try and understand InvoiceActions dataTypes.
